requirements.txt
streamlit
groq
Pillow